trunc nbinom/pois simulation cleanup, tests
